rename identity challenge dir to challenge
