+++ src/challenge/challenge-credential.cpp
