+++ src/challenge/challenge-credential.hpp
