+++ src/challenge/challenge-pin.hpp
