+++ tests/unit-tests/ca-module.t.cpp
@@ -19,9 +19,9 @@
  */
 
 #include "ca-module.hpp"
-#include "identity-challenge/challenge-module.hpp"
-#include "identity-challenge/challenge-email.hpp"
-#include "identity-challenge/challenge-pin.hpp"
+#include "challenge/challenge-module.hpp"
+#include "challenge/challenge-email.hpp"
+#include "challenge/challenge-pin.hpp"
 #include "detail/info-encoder.hpp"
 #include "requester.hpp"
 #include "test-common.hpp"
