+++ tests/unit-tests/challenge-pin.t.cpp
@@ -18,7 +18,7 @@
  * See AUTHORS.md for complete list of ndncert authors and contributors.
  */
 
-#include "identity-challenge/challenge-pin.hpp"
+#include "challenge/challenge-pin.hpp"
 #include "test-common.hpp"
 
 namespace ndn {
